+++ src/transcode.h
@@ -53,11 +53,11 @@ extern "C" {
 
 /* Functions for doing transcoding, called by main program body */
 struct transcoder* transcoder_new(char* filename);
-int transcoder_read(struct transcoder* trans, char* buff, int offset,
-                    int len);
+ssize_t transcoder_read(struct transcoder* trans, char* buff, off_t offset,
+                        size_t len);
 int transcoder_finish(struct transcoder* trans);
 void transcoder_delete(struct transcoder* trans);
-int transcoder_get_size(struct transcoder* trans);
+size_t transcoder_get_size(struct transcoder* trans);
 
 #ifdef __cplusplus
 }
